2020. 09. 29 <1> Ingame Resource 및 CineMachine 적용 Camera Size(10), Background Size(Scale 4) 임시 맞춤. Sorting Layer Argument 편성
<회의수정사항>
Gold Window 사이즈 조정 및 50% Transparent
Weapon Select Window의 잔여총알 표시 위치를 아이콘 안으로 넣고, 폰트 변경 및 Bold 해제
Move, Shot, Jump 버튼 시안성 좋은 Resource로 변경

+++ 2020_2_Project/Assets/Scripts/Objects/Player.cs
@@ -47,7 +47,7 @@ public class Player : MonoBehaviour
 
     public void Move(int _direction)
     {
-        _animator.SetInteger("direction", _direction);
+        _animator.SetFloat("direction", _direction);
         _animator.SetBool("move", true);
         _movePos.x = _direction;
         _directionPos.x = _direction;
@@ -63,7 +63,7 @@ public class Player : MonoBehaviour
     {
         if(_isLookup) // 위를 바라볼 때(키를 누르고 있을 때)
         {
-            _oldDirectionPos = _directionPos;
+            _oldDirectionPos = _directionPos; // 위를 바라보기 직전에 어느 시점을 보고 있었는지를 저장
             _directionPos = Vector2.up;
         }
         else // 위 그만 바라볼 때(키에서 손을 뗄 때)
